pyproject.toml
[project]
name = "sure"
version = "0.1.0"
description = "Add your description here"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "altair>=5.5.0",
  "boto3<1.36",
  "botocore<1.36",
  "celery[redis]>=5.5.3",
  "dataclasses-json>=0.6.7",
  "deepl>=1.25.0",
  "django>=5.2.6",
  "django-agent-trust>=1.1.0",
  "django-axes[ipware]>=8.0.0",
  "django-celery-beat>=2.8.1",
  "django-celery-results>=2.6.0",
  "django-colorfield>=0.14.0",
  "django-constance[redis]>=4.3.4",
  "django-cors-headers>=4.8.0",
  "django-crispy-forms>=2.5",
  "django-debug-toolbar>=6.0.0",
  "django-decorator-include>=3.3",
  "django-environ>=0.12.0",
  "django-health-check>=3.20.0",
  "django-model-utils>=5.0.0",
  "django-modeltranslation>=0.19.17",
  "django-ninja>=1.4.3",
  "django-otp>=1.6.3",
  "django-sesame>=3.2.3",
  "django-simple-history>=3.10.1",
  "django-storages>=1.14.6",
  "django-unfold>=0.67.0",
  "django-vite>=3.1.0",
  "fastexcel>=0.18.0",
  "gunicorn>=23.0.0",
  "openpyxl>=3.1.5",
  "pandas>=2.3.3",
  "phonenumbers>=9.0.15",
  "polars>=1.33.1",
  "psutil>=7.1.3",
  "psycopg2-binary>=2.9.10",
  "sentry-sdk[django]>=2.43.0",
  "twilio>=9.8.2",
  "whitenoise>=6.10.0",
  "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = [
  "autoflake>=2.3.1",
  "bandit>=1.8.6",
  "black>=25.1.0",
  "celery-types>=0.23.0",
  "django-stubs>=5.2.3",
  "isort>=6.0.1",
  "joblib-stubs>=1.5.2.0.20250831",
  "licensecheck>=2025.1.0",
  "mypy>=1.18.1",
  "pylint[spelling]>=3.3.8",
  "pyre-check>=0.9.21",
  "ruff>=0.13.1",
  "ty>=0.0.1a21",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"


[tool.bandit]
exclude = ["migrations", "tests", "./.venv/"]

[tool.pylint.MASTER]
ignore-paths = ['^.venv/.*$', '^migrations/.*$', '^tests/.*$', '^frontend/.*$']
disable = [
  "C0114",  # missing module docstring
  "C0115",  # missing class docstring
  "C0116",  # missing function or method docstring
  "R0903",  # too few public methods, for django models
  "R0901",  # too many ancestors
  "R6103",  # walrus operator
  "W0511",  # fixme
  "E1101", # no-member (false positive with django)
]
